+++ Source/Scene/WebMapServiceImageryProvider.js
@@ -114,10 +114,6 @@ define([
      * });
      *
      * viewer.scene.imageryLayers.addImageryProvider(provider);
-     *
-     * // This line is not required to use a WMS imagery layer, but adding it will enable automatic
-     * // display of WMS feature information (if available) on click.
-     * viewer.extend(Cesium.viewerEntityMixin);
      */
     var WebMapServiceImageryProvider = function WebMapServiceImageryProvider(options) {
         options = defaultValue(options, defaultValue.EMPTY_OBJECT);