+++ CHANGES.md
@@ -7,8 +7,9 @@ Change Log
   * Removed `GeometryPipeline.wrapLongitude`, which were previously deprecated in 1.4, use `GeometryPipeline.splitLongitude` instead.
   * Removed `GeometryPipeline.combine`, which were previously deprecated in 1.4, use `GeometryPipeline.combineInstances` instead.
 * Deprecated
-  * `Rectangle.intersectWith` was deprecated in Cesium 1.5. It will be removed in Cesium 1.6. Use `Rectangle.intersection`, which is the same but returns `undefined` when two rectangles do not intersect.
-  * `Rectangle.isEmpty` was deprecated in Cesium 1.5. It will be removed in Cesium 1.6.
+  * `Rectangle.intersectWith` was deprecated. It will be removed in Cesium 1.6. Use `Rectangle.intersection`, which is the same but returns `undefined` when two rectangles do not intersect.
+  * `Rectangle.isEmpty` was deprecated. It will be removed in Cesium 1.6.
+  * `viewerEntityMixin` was deprecated. It will be removed in Cesium 1.6. It's functionality has been made directly part of the `Viewer` widget.
 * Dramatically improve the quality of font outlines.
 * Improved polygon loading performance.
 * Fixed upsampleQuantizedTerrainMesh rounding errors that had occasionally led to missing terrain skirt geometry in upsampled tiles.
@@ -18,6 +19,9 @@ Change Log
 * Added `Math.mod` which computes `m % n` but also works when `m` is negative.
 * Fixed a bug in `PolylineGeometry` that would cause gaps in the line. [#2136](https://github.com/AnalyticalGraphicsInc/cesium/issues/2136)
 * Fixed imagery providers whose rectangle crosses the IDL. Added `Rectangle.computeWidth`, `Rectangle.computeHeight`, `Rectangle.width`, and `Rectangle.height`. [#2195](https://github.com/AnalyticalGraphicsInc/cesium/issues/2195)
+* Added `BoxGraphics` and `Entity.box`.
+* Added `CorridorGraphics` and `Entity.corridor`.
+* Added `CylinderGraphics` and `Entity.cylinder`.
 
 ### 1.4 - 2014-12-01